requirements.txt
fastapi
uvicorn
aiofiles
anthropic
requests
python-dotenv
beautifulsoup4
edgartools>=4.6.0
hishel>=0.1.3,<1.0
